requirements.txt
databases~=0.6.0
python-dotenv~=0.20.0
SQLAlchemy~=1.4.38
pydantic~=1.9.1
fastapi~=0.78.0
starlette~=0.19.1
alembic~=1.8.0
passlib~=1.7.4
bcrypt~=3.2.2